docs(func): homogenize punctuation on collections
It is better in my opinion to always use full punctuation, even in collections. This helps a lot keeping the documentation consistent, and not having to restrict the documentation writer to only **one** sentence to avoid having punctuation on some bullets and not on other bullets. The writing seems a bit more heavy at the first glance, but it ends up being more readable and consistent.