docs(tutorials appleminireemo): add details to help new users
As a developer with very low experience using MacOS, the path of eemo.ini files
would have saved me time while installing and configuring Reemo.
About the personal key, this is the vocabulary that is used on Reemo portal, which
is also confusing because Reemo uses **private key** on each one of their client
applications. So, having the two words clearly defined here, helps the user to
identify easily the Reemo actions.